feat(render): unify buffer management and implementation VMA

+++ src/Graphics/Resources/Components/TextureInfoComponent.hpp
@@ -2,5 +2,5 @@
 
 struct TextureInfoComponent
 {
-	size_t textureIndex;
+	int textureIndex;
 };
